+++ mtg/src/main/scala/com/github/gchudnov/mtg/internal/RelStatic.scala
@@ -1,8 +1,16 @@
 package com.github.gchudnov.mtg.internal
 
+import com.github.gchudnov.mtg.Mark
+import com.github.gchudnov.mtg.Domain
+import com.github.gchudnov.mtg.Interval
+
 /**
  * Static Interval Relations
  */
-private[mtg] transparent trait RelStatic {
-  // TODO: impl them
-}
+private[mtg] transparent trait RelStatic:
+
+  /**
+   * Relation: Before, Precedes (b)
+   */
+  final def before[T: Domain](a: Interval[T], b: Interval[T]): Boolean =
+    a.before(b)
